feat(core): add core extension
