+++ src/core/src/lib/extensions/core/core.ts
@@ -1,0 +1,121 @@
+import { FormlyExtension, FormlyConfig, TemplateManipulators } from '../../services/formly.config';
+import { FormlyFieldConfigCache, FormlyFieldConfig } from '../../components/formly.field.config';
+import { FormGroup, FormArray } from '@angular/forms';
+import { getFieldId, assignModelValue, isUndefined, clone, removeFieldControl, getFieldValue } from '../../utils';
+
+export class CoreExtension implements FormlyExtension {
+  private formId = 0;
+  constructor(private formlyConfig: FormlyConfig) { }
+
+  prePopulate(field: FormlyFieldConfigCache) {
+    if (field.parent) {
+      return;
+    }
+
+    const fieldTransforms = (field.options && field.options.fieldTransform) || this.formlyConfig.extras.fieldTransform;
+    (Array.isArray(fieldTransforms) ? fieldTransforms : [fieldTransforms]).forEach(fieldTransform => {
+      if (fieldTransform) {
+        const fieldGroup = fieldTransform(field.fieldGroup, field.model, <FormGroup>field.formControl, field.options);
+        if (!fieldGroup) {
+          throw new Error('fieldTransform must return an array of fields');
+        }
+      }
+    });
+  }
+
+  onPopulate(field: FormlyFieldConfigCache) {
+    this.initFieldOptions(field);
+    if (field.fieldGroup) {
+      field.fieldGroup.forEach((f, index) => {
+        Object.defineProperty(f, 'parent', { get: () => field, configurable: true });
+        Object.defineProperty(f, 'index', { get: () => index, configurable: true });
+        this.formId++;
+      });
+    }
+  }
+
+
+  private initFieldOptions(field: FormlyFieldConfigCache) {
+    const root = <FormlyFieldConfigCache> field.parent;
+    if (!root) {
+      return;
+    }
+
+    Object.defineProperty(field, 'options', { get: () => root.options, configurable: true });
+    Object.defineProperty(field, 'model', {
+      get: () => field.key && field.fieldGroup ? getFieldValue(field) : root.model,
+      configurable: true,
+    });
+
+    field.id = getFieldId(`formly_${this.formId}`, field, field['index']);
+    field.templateOptions = field.templateOptions || {};
+    field.modelOptions = field.modelOptions || {};
+    field.lifecycle = field.lifecycle || {};
+    if (field.type && field.key) {
+      field.templateOptions = Object.assign({
+        label: '',
+        placeholder: '',
+        focus: false,
+      }, field.templateOptions);
+    }
+
+    if (field.template && field.type !== 'formly-template') {
+      if (field.type) {
+        console.warn(`NgxFormly: passing 'type' property is not allowed when 'template' is set.`);
+      }
+      field.type = 'formly-template';
+    }
+
+    if (field.type) {
+      this.formlyConfig.getMergedField(field);
+    }
+    if (field.key && isUndefined(field.defaultValue) && (field.fieldGroup || field.fieldArray)) {
+      field.defaultValue = field.fieldArray ? [] : {};
+    }
+
+    if (!isUndefined(field.defaultValue) && isUndefined(getFieldValue(field))) {
+      assignModelValue(root.model, field.key, field.defaultValue);
+    }
+
+    this.initFieldWrappers(field);
+    if (field.fieldArray) {
+      this.initFieldArray(field);
+    }
+
+    if (!field.type && field.fieldGroup) {
+      field.type = 'formly-group';
+    }
+  }
+
+  private initFieldArray(field: FormlyFieldConfigCache) {
+    field.fieldGroup = field.fieldGroup || [];
+    if (field.fieldGroup.length > field.model.length) {
+      for (let i = field.fieldGroup.length; i >= field.model.length; --i) {
+        removeFieldControl(field.formControl as FormArray, i);
+        field.fieldGroup.splice(i, 1);
+      }
+    }
+
+    for (let i = field.fieldGroup.length; i < field.model.length; i++) {
+      const f = { ...clone(field.fieldArray), key: `${i}` };
+      field.fieldGroup.push(f);
+    }
+  }
+
+  private initFieldWrappers(field: FormlyFieldConfig) {
+    field.wrappers = field.wrappers || [];
+    const fieldTemplateManipulators: TemplateManipulators = {
+      preWrapper: [],
+      postWrapper: [],
+      ...(field.templateOptions.templateManipulators || {}),
+    };
+
+    field.wrappers = [
+      ...this.formlyConfig.templateManipulators.preWrapper.map(m => m(field)),
+      ...fieldTemplateManipulators.preWrapper.map(m => m(field)),
+      ...field.wrappers,
+      ...this.formlyConfig.templateManipulators.postWrapper.map(m => m(field)),
+      ...fieldTemplateManipulators.postWrapper.map(m => m(field)),
+    ].filter((el, i, a) => el && i === a.indexOf(el));
+  }
+}
