+++ src/core/src/lib/extensions/field-expression/field-expression.spec.ts
@@ -10,6 +10,7 @@ import { FormlyValueChangeEvent, FormlyFormOptionsCache } from '../../components
 import { FieldExpressionExtension } from './field-expression';
 import { FieldValidationExtension } from '../field-validation/field-validation';
 import { FieldFormExtension } from '../field-form/field-form';
+import { CoreExtension } from '../core/core';
 
 
 describe('FieldExpressionExtension', () => {
@@ -35,6 +36,7 @@ describe('FieldExpressionExtension', () => {
       wrappers: [{ name: 'label', component: TestComponent, types: ['input'] }],
       validators: [{ name: 'required', validation: Validators.required }],
       extensions: [
+        { name: 'core', extension: new CoreExtension(config) },
         { name: 'field-validation', extension: new FieldValidationExtension(config) },
         { name: 'field-form', extension: new FieldFormExtension() },
         { name: 'field-expression', extension: new FieldExpressionExtension() },
